Move code calling BIOS from bioshd.c to bios.c
