+++ elks/arch/i86/drivers/block/Makefile
@@ -37,7 +37,7 @@ OBJS  = init.o genhd.o ll_rw_blk.o
 
 # BIOS floppy and hard drive support
 ifeq ($(CONFIG_BLK_DEV_BIOS), y)
-OBJS += bioshd.o idequery.o
+OBJS += bioshd.o bios.o idequery.o
 endif
 
 ifeq ($(CONFIG_BLK_DEV_RAM), y)
