+++ elks/arch/i86/drivers/block/bioshd.c
@@ -64,27 +64,23 @@
 /* the following must match with /dev minor numbering scheme*/
 #define NUM_MINOR       32      /* max minor devices per drive*/
 #define MINOR_SHIFT     5       /* =log2(NUM_MINOR) shift to get drive num*/
-#define NUM_DRIVES      8       /* =256/NUM_MINOR max number of drives*/
-#define DRIVE_FD0       4       /* first floppy drive =NUM_DRIVES/2*/
-#define DRIVE_FD1       5       /* second floppy drive*/
-#define DRIVE_FD2       6       /* PC98 only*/
-#define DRIVE_FD3       7       /* PC98 only*/
+#define DRIVE_HD0       0
+#define DRIVE_FD0       4       /* =NUM_DRIVES/2 first floppy drive*/
 
-#define MAJOR_NR BIOSHD_MAJOR
-#define BIOSDISK
-
-#include "blk.h"
-
-#ifdef CONFIG_ARCH_IBMPC
-#define MAXDRIVES       2       /* max floppy drives*/
-#endif
-
+#define HD_DRIVES       4       /* max hard drives */
 #ifdef CONFIG_ARCH_PC98
-#define MAXDRIVES       4       /* max floppy drives*/
+#define FD_DRIVES       4
+#else
+#define FD_DRIVES       2       /* max floppy drives */
 #endif
+#define NUM_DRIVES      (HD_DRIVES+FD_DRIVES) /* max number of drives (<=256/NUM_MINOR) */
 
 /* comment out following line for single-line drive info summary*/
 #define PRINT_DRIVE_INFO        NUM_DRIVES
+
+#define MAJOR_NR        BIOSHD_MAJOR
+#define BIOSDISK
+#include "blk.h"
 
 struct elks_disk_parms {
     __u16 track_max;            /* number of tracks, little-endian */
@@ -96,461 +92,54 @@ struct elks_disk_parms {
 } __attribute__((packed));
 
 static int bioshd_initialized = 0;
-static struct biosparms bdt;
-
-/* Useful defines for accessing the above structure. */
-#define CARRY_SET (bdt.fl & 0x1)
-#define BD_AX bdt.ax
-#define BD_BX bdt.bx
-#define BD_CX bdt.cx
-#define BD_DX bdt.dx
-#define BD_SI bdt.si
-#define BD_DI bdt.di
-#define BD_BP bdt.bp
-#define BD_ES bdt.es
-#define BD_FL bdt.fl
-
-static struct drive_infot drive_info[NUM_DRIVES];
-
-static struct hd_struct hd[NUM_DRIVES << MINOR_SHIFT];  /* partitions start, size*/
-static int hd_sizes[NUM_DRIVES << MINOR_SHIFT];         /* used only with BDEV_SIZE_CHK*/
-
-static int access_count[NUM_DRIVES];    /* for invalidating buffers/inodes*/
-
 static int fd_count = 0;                /* number of floppy disks */
 static int hd_count = 0;                /* number of hard disks */
 
-struct drive_infot fd_types[] = {       /* AT/PS2 BIOS reported floppy formats*/
-    {40,  9, 2, 512, 0},
-    {80, 15, 2, 512, 1},
-    {80,  9, 2, 512, 2},
-    {80, 18, 2, 512, 3},
-    {80, 36, 2, 512, 4},
-#ifdef CONFIG_ARCH_PC98
-    {77,  8, 2, 1024,5},
-#endif
-};
+static int access_count[NUM_DRIVES];    /* device open count */
+static struct drive_infot drive_info[NUM_DRIVES];   /* operating drive info */
+static struct drive_infot *cache_drive;
+struct drive_infot *last_drive;         /* set to last drivep-> used in read/write */
+extern struct drive_infot fd_types[];   /* BIOS floppy formats */
 
-#ifdef CONFIG_ARCH_PC98
-unsigned char hd_drive_map[NUM_DRIVES] = {/* BIOS drive mappings*/
-    0xA0, 0xA1, 0xA2, 0xA3,             /* hda, hdb */
-#ifdef CONFIG_IMG_FD1232
-    0x90, 0x91, 0x92, 0x93              /* fd0, fd1 */
-#else
-    0x30, 0x31, 0x32, 0x33              /* fd0, fd1 */
-#endif
-};
-#else
-unsigned char hd_drive_map[NUM_DRIVES] = {/* BIOS drive mappings*/
-    0x80, 0x81, 0x82, 0x83,             /* hda, hdb */
-    0x00, 0x01, 0x02, 0x03              /* fd0, fd1 */
-};
-#endif
+static struct hd_struct hd[NUM_DRIVES << MINOR_SHIFT];  /* partitions start, size*/
+//static int hd_sizes[NUM_DRIVES << MINOR_SHIFT];       /* used only with BDEV_SIZE_CHK*/
 
-#define SPT             4       /* DDPT offset of sectors per track*/
-static unsigned char DDPT[14];  /* our copy of diskette drive parameter table*/
-unsigned long __far *vec1E = _MK_FP(0, 0x1E << 2);
-
-static int bioshd_ioctl(struct inode *, struct file *, unsigned int, unsigned int);
 static int bioshd_open(struct inode *, struct file *);
 static void bioshd_release(struct inode *, struct file *);
+static int bioshd_ioctl(struct inode *, struct file *, unsigned int, unsigned int);
 static void bioshd_geninit(void);
-static void set_cache_invalid(void);
 
 static struct gendisk bioshd_gendisk = {
     MAJOR_NR,                   /* Major number */
     "hd",                       /* Major name */
     MINOR_SHIFT,                /* Bits to shift to get real from partition */
     1 << MINOR_SHIFT,           /* Number of partitions per real */
-    NUM_DRIVES,                 /* maximum number of real */
+    NUM_DRIVES,                 /* maximum number of drives */
     bioshd_geninit,             /* init function */
     hd,                         /* hd struct */
-    hd_sizes,                   /* sizes not blocksizes */
-    0,                          /* number */
-    (void *) drive_info,        /* internal */
+    0,//hd_sizes,               /* sizes not blocksizes */
+    0,                          /* hd drives found */
+    drive_info,
     NULL                        /* next */
 };
 
-struct drive_infot *last_drive; /* set to last drivep-> used in read/write */
-static struct drive_infot *cache_drive;
-
 static void set_cache_invalid(void)
 {
-        cache_drive = NULL;
+    cache_drive = NULL;
 }
-
-/* As far as I can tell this doesn't actually work, but we might
- * as well try it -- Some XT controllers are happy with it.. [AC]
- */
-
-static void bios_disk_reset(int drive)
-{
-#ifdef CONFIG_ARCH_PC98
-    BD_AX = BIOSHD_RESET | drive;
-#else
-    BD_AX = BIOSHD_RESET;
-    BD_DX = drive;
-#endif
-    call_bios(&bdt);
-    /* ignore errors with carry set*/
-}
-
-static int bios_disk_rw(unsigned cmd, unsigned num_sectors, unsigned drive,
-        unsigned cylinder, unsigned head, unsigned sector, unsigned seg, unsigned offset)
-{
-#ifdef CONFIG_ARCH_PC98
-    BD_AX = cmd | drive;
-    if (((0xF0 & drive) == 0x80) || ((0xF0 & drive) == 0xA0)) {
-        BD_BX = (unsigned int) (num_sectors << 9);
-        BD_CX = cylinder;
-        BD_DX = (head << 8) | ((sector - 1) & 0xFF);
-    }
-    else {
-        if ((0xF0 & drive) == 0x90) {
-            BD_BX = (unsigned int) (num_sectors << 10);
-            BD_CX = (3 << 8) | cylinder;
-        }
-        else {
-            BD_BX = (unsigned int) (num_sectors << 9);
-            BD_CX = (2 << 8) | cylinder;
-        }
-        BD_DX = (head << 8) | sector;
-    }
-    BD_ES = seg;
-    BD_BP = offset;
-#else
-
-#if RESET_DISK_CHG
-    static unsigned last = 0;
-    if (drive != last) {
-        bios_disk_reset(1); /* fixes QEMU retry when switching drive types #1119 */
-        last = drive;
-    }
-#endif
-    BD_AX = cmd | num_sectors;
-    BD_CX = (unsigned int) ((cylinder << 8) | ((cylinder >> 2) & 0xc0) | sector);
-    BD_DX = (head << 8) | drive;
-    BD_ES = seg;
-    BD_BX = offset;
-#endif
-    debug_bios("BIOSHD(%x): %s CHS %d/%d/%d count %d\n", drive,
-        cmd==BIOSHD_READ? "read": "write",
-        cylinder, head, sector, num_sectors);
-#ifdef IODELAY
-    /* emulate floppy delay for QEMU */
-    unsigned long timeout = jiffies + IODELAY*HZ/100;
-    while (!time_after(jiffies, timeout)) continue;
-#endif
-    return call_bios(&bdt);
-}
-
-#ifdef CONFIG_BLK_DEV_BHD
-/* This function checks to see which hard drives are active and sets up the
- * drive_info[] table for them.  Ack this is darned confusing...
- */
-static unsigned short int INITPROC bios_gethdinfo(void) {
-    int drive, ndrives = 0;
-    register struct drive_infot *drivep = &drive_info[0];
-
-#ifdef CONFIG_ARCH_PC98
-    int ide_drives = 0;
-    int scsi_id;
-    int device_type;
-    int call_bios_rvalue;
-
-    /* IDE */
-    for (drive = 0; drive < 4; drive++) {
-        if (peekb(0x55D,0) & (1 << drive)) {
-            BD_AX = BIOSHD_MODESET | (drive + 0x80);
-            BD_ES = BD_DI = BD_SI = 0;
-            call_bios(&bdt);
-            hd_drive_map[ide_drives++] = drive + 0x80;
-        }
-    }
-    if (ide_drives > 0)
-        printk("bioshd: Detected IDE hd.\n");
-    ndrives = ide_drives;
-
-    /* SCSI */
-    if (ndrives < 4) {
-        for (scsi_id = 0; scsi_id < 7; scsi_id++) {
-            BD_AX = BIOSHD_DRIVE_PARMS | (scsi_id + 0xA0);
-            BD_ES = BD_DI = BD_SI = 0;
-            call_bios_rvalue = call_bios(&bdt);
-            if ((call_bios_rvalue == 0) && (BD_DX & 0xff)) {
-                BD_AX = BIOSHD_DEVICE_TYPE | (scsi_id + 0xA0);
-                BD_ES = BD_DI = BD_SI = 0;
-                BD_BX = 0;
-                call_bios(&bdt);
-                device_type = BD_BX & 0xf; /* device_type = 0 for Hard Disk */
-                if (device_type == 0)
-                    hd_drive_map[ndrives++] = scsi_id + 0xA0;
-            }
-            if (ndrives >= 4) break;
-        }
-    }
-    if (ndrives > ide_drives)
-        printk("bioshd: Detected SCSI hd.\n");
-#else
-    BD_AX = BIOSHD_DRIVE_PARMS;
-    BD_DX = 0x80;               /* query hard drives only*/
-    BD_ES = BD_DI = BD_SI = 0;  /* guard against BIOS bugs*/
-    if (!call_bios(&bdt))
-        ndrives = BD_DX & 0xff;
-    else
-        debug_bios("bioshd: get_drive_parms fail on hd\n");
-#endif
-    if (ndrives > NUM_DRIVES/2)
-        ndrives = NUM_DRIVES/2;
-
-    for (drive = 0; drive < ndrives; drive++) {
-#ifdef CONFIG_ARCH_PC98
-        BD_AX = BIOSHD_DRIVE_PARMS | hd_drive_map[drive];
-#else
-        BD_AX = BIOSHD_DRIVE_PARMS;
-        BD_DX = drive + 0x80;
-#endif
-        BD_ES = BD_DI = BD_SI = 0;      /* guard against BIOS bugs*/
-        if (call_bios(&bdt) == 0) {
-#ifdef CONFIG_ARCH_PC98
-            drivep->heads = BD_DX >> 8;
-            drivep->sectors = BD_DX & 0xff;
-            drivep->cylinders = BD_CX;
-#else
-            drivep->heads = (BD_DX >> 8) + 1;
-            drivep->sectors = BD_CX & 0x3f;
-            /* NOTE: some BIOS may underreport cylinders by 1*/
-            drivep->cylinders = (((BD_CX & 0xc0) << 2) | (BD_CX >> 8)) + 1;
-#endif
-            drivep->fdtype = -1;
-            drivep->sector_size = 512;
-            printk("bioshd: hd%c BIOS CHS %u,%d,%d\n", 'a'+drive, drivep->cylinders,
-                drivep->heads, drivep->sectors);
-        }
-#ifdef CONFIG_IDE_PROBE
-        if (sys_caps & CAP_HD_IDE) {            /* Normally PC/AT or higher */
-            if (!get_ide_data(drive, drivep)) { /* get CHS from the drive itself */
-                /* sanity checks already done, accepting data */
-                printk("bioshd: hd%c  IDE CHS %d,%d,%d\n", 'a'+drive, drivep->cylinders,
-                drivep->heads, drivep->sectors);
-            }
-        }
-#endif
-        drivep++;
-    }
-    return ndrives;
-}
-#endif
-
-#ifdef CONFIG_BLK_DEV_BFD_HARD
-/* hard-coded floppy configuration*/
-static unsigned short int INITPROC bios_getfdinfo(void)
-{
-/* Set this to match your system. Currently it's set to a two drive system:
- *
- *              720KB as /dev/fd0
- *      and     720KB as /dev/fd1
- *
- * ndrives is number of drives in your system (either 0, 1 or 2)
- */
-
-    int ndrives = MAXDRIVES;
-
-/* drive_info[] should be set *only* for existing drives;
- * comment out drive_info lines if you don't need them
- * (e.g. you have less than 2 drives)
- *
- * Enter type 4 in fd_types' brackets for unknown drive type
- * Otherwise use floppy drive type table below:
- *
- *      Type    Format
- *      ~~~~    ~~~~~~
- *        0     360 KB
- *        1     1.2 MB
- *        2     720 KB
- *        3     1.44 MB
- *        4     2.88 MB or Unknown
- *        5     1.232 MB (PC98 1K sectors)
- *
- * Warning: drive will be reported as 2880 KB at bootup if you've set it
- * as unknown (4). Floppy probe will detect correct floppy format at each
- * change so don't bother with that
- */
-
-#ifdef CONFIG_ARCH_PC98
-#if defined(CONFIG_IMG_FD1232)
-    drive_info[DRIVE_FD0] = fd_types[5];
-    drive_info[DRIVE_FD1] = fd_types[5];
-    drive_info[DRIVE_FD2] = fd_types[5];
-    drive_info[DRIVE_FD3] = fd_types[5];
-#elif defined(CONFIG_IMG_FD1440)
-    drive_info[DRIVE_FD0] = fd_types[3];
-    drive_info[DRIVE_FD1] = fd_types[3];
-    drive_info[DRIVE_FD2] = fd_types[3];
-    drive_info[DRIVE_FD3] = fd_types[3];
-#endif
-#endif
-
-#ifdef CONFIG_ARCH_IBMPC
-    drive_info[DRIVE_FD0] = fd_types[2];        /*  /dev/fd0    */
-    drive_info[DRIVE_FD1] = fd_types[2];        /*  /dev/fd1    */
-#endif
-
-    return ndrives;
-}
-
-#elif defined(CONFIG_BLK_DEV_BFD)
-
-/* use BIOS to query floppy configuration*/
-static unsigned short int INITPROC bios_getfdinfo(void)
-{
-    register struct drive_infot *drivep = &drive_info[DRIVE_FD0];
-    int drive, ndrives = 0;
-
-#ifndef CONFIG_ROMCODE
-    /*
-     * The INT 13h floppy query will fail on IBM XT v1 BIOS and earlier,
-     * so default to # drives from the BIOS data area at 0x040:0x0010 (INT 11h).
-     */
-    unsigned char equip_flags = peekb(0x10, 0x40);
-    if (equip_flags & 0x01)
-        ndrives = (equip_flags >> 6) + 1;
-#endif
-
-#ifdef CONFIG_ARCH_PC98
-    for (drive = 0; drive < 4; drive++) {
-        if (peekb(0x55C,0) & (1 << drive)) {
-#ifdef CONFIG_IMG_FD1232
-            hd_drive_map[DRIVE_FD0 + ndrives] = drive + 0x90;
-            *drivep = fd_types[5];
-#else
-            hd_drive_map[DRIVE_FD0 + ndrives] = drive + 0x30;
-            *drivep = fd_types[3];
-#endif
-            ndrives++;  /* floppy drive count*/
-            drivep++;
-        }
-    }
-#else
-    /* Floppy query may fail if not PC/AT */
-    BD_AX = BIOSHD_DRIVE_PARMS;
-    BD_DX = 0;                          /* query floppies only*/
-    BD_ES = BD_DI = BD_SI = 0;          /* guard against BIOS bugs*/
-    if (!call_bios(&bdt)) {
-        int drives = BD_DX & 0xff;      /* floppy drive count */
-        if (!drives && ndrives) {       /* handle Toshiba T1100 BIOS returning 0 drives */
-            for (drive = 0; drive < ndrives; drive++) {
-                printk("fd%d: default 720k\n", drive);
-                *drivep++ = fd_types[2];
-            }
-            return ndrives;
-        } else ndrives = drives;
-    } else
-        printk("fd: no get drive fn, ndrives %d\n", ndrives);
-
-    /* set drive type for floppies*/
-    for (drive = 0; drive < ndrives; drive++) {
-        /*
-         * If type cannot be determined using BIOSHD_DRIVE_PARMS,
-         * set drive type to 1.4MM on AT systems, and 360K for XT.
-         */
-        BD_AX = BIOSHD_DRIVE_PARMS;
-        BD_DX = drive;
-        BD_ES = BD_DI = BD_SI = 0;      /* guard against BIOS bugs*/
-        if (!call_bios(&bdt))           /* returns drive type in BL*/
-            *drivep = fd_types[(BD_BX & 0xFF) - 1];
-        else {
-            int type = (sys_caps & CAP_PC_AT) ? 3 : 0;
-            *drivep = fd_types[type];
-            printk("fd%d: default %s\n", drive, type? "1440k": "360k");
-        }
-        drivep++;
-    }
-#endif
-    return ndrives;
-}
-#endif
-
-static void bioshd_release(struct inode *inode, struct file *filp)
-{
-    kdev_t dev = inode->i_rdev;
-    int target = DEVICE_NR(dev);
-
-    if (--access_count[target] == 0) {
-        fsync_dev(dev);
-        invalidate_inodes(dev);
-        invalidate_buffers(dev);
-    }
-}
-
-/* set our DDPT sectors per track value*/
-static void set_ddpt(int max_sectors)
-{
-        DDPT[SPT] = (unsigned char) max_sectors;
-}
-
-/* get the diskette drive parameter table from INT 1E and point to our RAM copy of it*/
-static void copy_ddpt(void)
-{
-        unsigned long oldvec = *vec1E;
-
-        /* We want to prevent the BIOS from accidentally doing a "multitrack"
-         * floppy read --- and wrapping around from one head to the next ---
-         * when ELKS only wants to read from a single track.
-         *
-         * (E.g. if DDPT SPT = 9, and a disk has 18 sectors per track, and we
-         * want to read sectors 9--10 from track 0, side 0, then the BIOS may
-         * read sector 9 from track 0, side 0, followed by sector 1 from track
-         * 0, side 1, which will be wrong.)
-         *
-         * To prevent this, we set the DDPT SPT field to the actual sector
-         * count per track in the detected disk geometry.  The DDPT SPT
-         * should never be smaller than the actual SPT, but it can be larger.
-         *
-         * Rather than issue INT 13h function 8 (Get Drive Parameters, not implemented
-         * on IBM XT BIOS v1 and earlier) to get an accurate DDPT, just copy the original
-         * DDPT to RAM, where the sectors per track value will be modified before each
-         * INT 13h function 2/3 (Read/Write Disk Sectors).
-         * Using a patched DDPT also eliminates the need for a seperate fix for #39/#44.
-         */
-        fmemcpyw(DDPT, _FP_SEG(DDPT), (void *)(unsigned)oldvec, _FP_SEG(oldvec),
-                sizeof(DDPT)/2);
-        debug_bios("bioshd: DDPT vector %x:%x SPT %d\n", _FP_SEG(oldvec),
-            (unsigned)oldvec, DDPT[SPT]);
-        *vec1E = (unsigned long)(void __far *)DDPT;
-}
-
-/* map drives */
-static void map_drive(int *drive)
-{
-        *drive = hd_drive_map[*drive];
-}
-
-#ifdef CONFIG_ARCH_PC98
-/* switch device */
-static void switch_device98(int target, unsigned char device, struct drive_infot *drivep)
-{
-    hd_drive_map[target + DRIVE_FD0] = (device | (hd_drive_map[target + DRIVE_FD0] & 0x0F));
-    if (device == 0x30)
-        *drivep = fd_types[3];  /* 1.44 MB */
-    else if (device == 0x90)
-        *drivep = fd_types[5];  /* 1.232 MB */
-}
-#endif
 
 static int read_sector(int drive, int cylinder, int sector)
 {
     int count = 2;              /* one retry on probe or boot sector read */
 
 #ifdef CONFIG_ARCH_PC98
-    drive += DRIVE_FD0;
-    map_drive(&drive);
+    drive = bios_drive_map[DRIVE_FD0+drive];
 #endif
 
     set_cache_invalid();
     do {
         set_irq();
-        set_ddpt(36);           /* set to large value to avoid BIOS issues*/
+        bios_set_ddpt(36);      /* set to large value to avoid BIOS issues*/
         if (!bios_disk_rw(BIOSHD_READ, 1, drive, cylinder, 0, sector, DMASEG, 0))
             return 0;           /* everything is OK */
         bios_disk_reset(drive);
@@ -582,7 +171,7 @@ static void probe_floppy(int target, struct hd_struct *hdp)
 #endif
         int count, found_PB = 0;
 
-        target &= MAXDRIVES - 1;
+        target &= FD_DRIVES - 1;
 
 #if !FORCE_PROBE
         /* Try to look for an ELKS or DOS parameter block in the first sector.
@@ -660,10 +249,10 @@ static void probe_floppy(int target, struct hd_struct *hdp)
 #ifdef CONFIG_ARCH_PC98
         do {
             if (count)
-                switch_device98(target, 0x30, drivep);  /* 1.44 MB */
+                bios_switch_device98(target, 0x30, drivep);  /* 1.44 MB */
             /* skip probing first entry */
             if (count && read_sector(target, track_probe[count] - 1, 1)) {
-                switch_device98(target, 0x90, drivep);  /* 1.232 MB */
+                bios_switch_device98(target, 0x90, drivep);  /* 1.232 MB */
                 break;
             }
             drivep->cylinders = track_probe[count];
@@ -693,10 +282,10 @@ static void probe_floppy(int target, struct hd_struct *hdp)
 #ifdef CONFIG_ARCH_PC98
         do {
             if (count)
-                switch_device98(target, 0x30, drivep);  /* 1.44 MB */
+                bios_switch_device98(target, 0x30, drivep);  /* 1.44 MB */
             /* skip reading first entry */
             if (count && read_sector(target, 0, sector_probe[count])) {
-                switch_device98(target, 0x90, drivep);  /* 1.232 MB */
+                bios_switch_device98(target, 0x90, drivep);  /* 1.232 MB */
                 break;
             }
             drivep->sectors = sector_probe[count];
@@ -760,6 +349,18 @@ static int bioshd_open(struct inode *inode, struct file *filp)
     return 0;
 }
 
+static void bioshd_release(struct inode *inode, struct file *filp)
+{
+    kdev_t dev = inode->i_rdev;
+    int target = DEVICE_NR(dev);
+
+    if (--access_count[target] == 0) {
+        fsync_dev(dev);
+        invalidate_inodes(dev);
+        invalidate_buffers(dev);
+    }
+}
+
 static struct file_operations bioshd_fops = {
     NULL,                       /* lseek - default */
     block_read,                 /* read - general block-dev read */
@@ -780,11 +381,11 @@ void INITPROC bioshd_init(void)
     outb_p(0x0C, FDC_DOR);      /* FD motors off, enable IRQ and DMA*/
 
 #ifdef CONFIG_BLK_DEV_BFD
-    fd_count = bios_getfdinfo();
+    fd_count = bios_getfdinfo(&drive_info[DRIVE_FD0]);
 #endif
 #ifdef CONFIG_BLK_DEV_BHD
-    hd_count = bios_gethdinfo();
-    bioshd_gendisk.nr_real = hd_count;
+    hd_count = bios_gethdinfo(&drive_info[DRIVE_HD0]);
+    bioshd_gendisk.nr_hd = hd_count;
 #endif
 
 #ifdef PRINT_DRIVE_INFO
@@ -836,7 +437,7 @@ void INITPROC bioshd_init(void)
 
     if (!(fd_count + hd_count)) return;
 
-    copy_ddpt();        /* make a RAM copy of the disk drive parameter table*/
+    bios_copy_ddpt();       /* make a RAM copy of the disk drive parameter table*/
 
     if (!register_blkdev(MAJOR_NR, DEVICE_NAME, &bioshd_fops)) {
         blk_dev[MAJOR_NR].request_fn = DEVICE_REQUEST;
@@ -846,7 +447,7 @@ void INITPROC bioshd_init(void)
             gendisk_head = &bioshd_gendisk;
         } else {
             for (ptr = gendisk_head; ptr->next != NULL; ptr = ptr->next)
-                /* Do nothing */ ;
+                continue;
             ptr->next = &bioshd_gendisk;
             //bioshd_gendisk.next = NULL;
         }
@@ -854,6 +455,32 @@ void INITPROC bioshd_init(void)
     } else {
         printk("bioshd: init error\n");
     }
+}
+
+static void INITPROC bioshd_geninit2(void)
+{
+    struct drive_infot *drivep = drive_info;
+    struct hd_struct *hdp = hd;
+    int i;
+
+    for (i = 0; i < NUM_DRIVES << MINOR_SHIFT; i++) {
+        if ((i & ((1 << MINOR_SHIFT) - 1)) == 0) {
+            hdp->nr_sects = (sector_t)drivep->sectors * drivep->heads * drivep->cylinders;
+            hdp->start_sect = 0;
+            drivep++;
+        } else {
+            hdp->nr_sects = 0;
+            hdp->start_sect = -1;
+        }
+        hdp++;
+    }
+
+}
+
+/* called by setup_dev() */
+static void bioshd_geninit(void)
+{
+    bioshd_geninit2();
 }
 
 static int bioshd_ioctl(struct inode *inode, struct file *file, unsigned int cmd,
@@ -930,8 +557,7 @@ static int do_readwrite(struct drive_infot *drivep, sector_t start, char *buf,
     size_t end;
     int usedmaseg;
 
-    drive = drivep - drive_info;
-    map_drive(&drive);
+    drive = bios_drive_map[drivep - drive_info];
     get_chst(drivep, &start, &cylinder, &head, &sector, &this_pass, 0);
 
     /* limit I/O to requested sector count*/
@@ -964,7 +590,7 @@ static int do_readwrite(struct drive_infot *drivep, sector_t start, char *buf,
         debug_bios("bioshd(%x): cmd %d CHS %d/%d/%d count %d\n",
             drive, cmd, cylinder, head, sector, this_pass);
 
-        set_ddpt(drivep->sectors);
+        bios_set_ddpt(drivep->sectors);
         error = bios_disk_rw(cmd == WRITE? BIOSHD_WRITE: BIOSHD_READ, this_pass,
                                 drive, cylinder, head, sector, segment, offset);
         if (error) {
@@ -996,7 +622,7 @@ static void do_readtrack(struct drive_infot *drivep, sector_t start)
     int drive = drivep - drive_info;
     int error, errs = 0;
 
-    map_drive(&drive);
+    drive = bios_drive_map[drive];
     get_chst(drivep, &start, &cylinder, &head, &sector, &num_sectors, 1);
 
     if (num_sectors > (DMASEGSZ / drivep->sector_size))
@@ -1006,7 +632,7 @@ static void do_readtrack(struct drive_infot *drivep, sector_t start)
         debug_bios("bioshd(%x): track read CHS %d/%d/%d count %d\n",
                 drive, cylinder, head, sector, num_sectors);
 
-        set_ddpt(drivep->sectors);
+        bios_set_ddpt(drivep->sectors);
         error = bios_disk_rw(BIOSHD_READ, num_sectors, drive,
                                  cylinder, head, sector, DMASEG, 0);
         if (error) {
@@ -1039,13 +665,14 @@ static int cache_valid(struct drive_infot *drivep, sector_t start, char *buf,
         return 0;
 
     offset = (int)(start - cache_startsector) * drivep->sector_size;
-    debug_bios("bioshd(%x): cache hit lba %ld\n", hd_drive_map[drivep-drive_info], start);
+    debug_bios("bioshd(%x): cache hit lba %ld\n",
+        bios_drive_map[drivep-drive_info], start);
     xms_fmemcpyw(buf, seg, (void *)offset, DMASEG, drivep->sector_size >> 1);
     return 1;
 }
 
-static int cache_tries = 0;
-static int cache_hits = 0;
+static int cache_tries;
+static int cache_hits;
 
 /* read from cache, return # sectors read*/
 static int do_cache_read(struct drive_infot *drivep, sector_t start, char *buf,
@@ -1093,7 +720,7 @@ next_block:
         drivep = &drive_info[drive];
 
         /* make sure it's a disk that we are dealing with. */
-        if (drive > (DRIVE_FD0 + MAXDRIVES - 1) || drivep->heads == 0) {
+        if (drive > (DRIVE_FD0 + FD_DRIVES - 1) || drivep->heads == 0) {
             printk("bioshd: non-existent drive\n");
             end_request(0);
             continue;
@@ -1196,56 +823,3 @@ static int revalidate_hddisk(int dev, int maxusage)
     return 0;
 }
 #endif
-
-static void bioshd_geninit(void)
-{
-    register struct drive_infot *drivep;
-    register struct hd_struct *hdp = hd;
-    int i;
-
-    drivep = drive_info;
-    for (i = 0; i < NUM_DRIVES << MINOR_SHIFT; i++) {
-        if ((i & ((1 << MINOR_SHIFT) - 1)) == 0) {
-            hdp->nr_sects = (sector_t) drivep->sectors *
-                drivep->heads * drivep->cylinders;
-            hdp->start_sect = 0;
-            drivep++;
-        } else {
-            hdp->nr_sects = 0;
-            hdp->start_sect = -1;
-        }
-        hdp++;
-    }
-
-}
-
-/* convert a bios drive number to a bioshd dev_t*/
-dev_t INITPROC bios_conv_bios_drive(unsigned int biosdrive)
-{
-    int minor;
-    int partition = 0;
-    extern int boot_partition;
-
-#ifdef CONFIG_ARCH_PC98
-    if (((biosdrive & 0xF0) == 0x80) || ((biosdrive & 0xF0) == 0xA0)) { /* hard drive*/
-        for (minor = 0; minor < 4; minor++) {
-            if (biosdrive == hd_drive_map[minor]) break;
-        }
-        if (minor >= 4) minor = 0;
-        partition = boot_partition;     /* saved from add_partition()*/
-    } else {
-        for (minor = 4; minor < 8; minor++) {
-            if (biosdrive == hd_drive_map[minor]) break;
-        }
-        if (minor >= 8) minor = 4;
-    }
-#else
-    if (biosdrive & 0x80) {             /* hard drive*/
-        minor = biosdrive & 0x03;
-        partition = boot_partition;     /* saved from add_partition()*/
-    } else
-        minor = (biosdrive & 0x03) + DRIVE_FD0;
-#endif
-
-    return MKDEV(BIOSHD_MAJOR, (minor << MINOR_SHIFT) + partition);
-}
