+++ elks/include/linuxmt/genhd.h
@@ -4,11 +4,11 @@
 #include <linuxmt/types.h>
 
 /*
- * 	genhd.h Copyright (C) 1992 Drew Eckhardt
- *	Generic hard disk header file by  
- * 		Drew Eckhardt
+ *      genhd.h Copyright (C) 1992 Drew Eckhardt
+ *      Generic hard disk header file by
+ *              Drew Eckhardt
  *
- *		<drew@colorado.edu>
+ *              <drew@colorado.edu>
  */
 
 #define CONFIG_MSDOS_PARTITION 1
@@ -21,33 +21,33 @@
 
 struct partition
 {
-    unsigned char boot_ind;	/* 0x80 - active */
-    unsigned char head;		/* starting head */
-    unsigned char sector;	/* starting sector */
-    unsigned char cyl;		/* starting cylinder */
-    unsigned char sys_ind;	/* What partition type */
-    unsigned char end_head;	/* end head */
-    unsigned char end_sector;	/* end sector */
-    unsigned char end_cyl;	/* end cylinder */
-    sector_t start_sect;	/* starting sector counting from 0 */
-    sector_t nr_sects;		/* nr of sectors in partition */
+    unsigned char boot_ind;     /* 0x80 - active */
+    unsigned char head;         /* starting head */
+    unsigned char sector;       /* starting sector */
+    unsigned char cyl;          /* starting cylinder */
+    unsigned char sys_ind;      /* What partition type */
+    unsigned char end_head;     /* end head */
+    unsigned char end_sector;   /* end sector */
+    unsigned char end_cyl;      /* end cylinder */
+    sector_t start_sect;        /* starting sector counting from 0 */
+    sector_t nr_sects;          /* nr of sectors in partition */
 };
 
 #ifdef CONFIG_ARCH_PC98
 struct partition_pc98
 {
-    unsigned char boot_ind;	/* bootable */
-    unsigned char active;	/* active or sleep */
+    unsigned char boot_ind;     /* bootable */
+    unsigned char active;       /* active or sleep */
     unsigned int dummy;
-    unsigned char ipl_sector;	/* ipl sector */
-    unsigned char ipl_head;	/* ipl head */
-    unsigned int ipl_cyl;  	/* ipl cylinder */
-    unsigned char sector;	/* starting sector */
-    unsigned char head;		/* starting head */
-    unsigned int cyl;	  	/* starting cylinder */
-    unsigned char end_sector;	/* end sector */
-    unsigned char end_head;	/* end head */
-    unsigned int end_cyl;	/* end cylinder */
+    unsigned char ipl_sector;   /* ipl sector */
+    unsigned char ipl_head;     /* ipl head */
+    unsigned int ipl_cyl;       /* ipl cylinder */
+    unsigned char sector;       /* starting sector */
+    unsigned char head;         /* starting head */
+    unsigned int cyl;           /* starting cylinder */
+    unsigned char end_sector;   /* end sector */
+    unsigned char end_head;     /* end head */
+    unsigned int end_cyl;       /* end cylinder */
     char name[16];
 };
 #endif
@@ -58,24 +58,32 @@ struct hd_struct
     sector_t nr_sects;
 };
 
+struct drive_infot {            /* CHS per drive*/
+    unsigned int cylinders;
+    int sectors;
+    int heads;
+    int sector_size;
+    int fdtype;                 /* floppy fd_types[] index  or -1 if hd */
+};
+
 struct gendisk
 {
-    int major;			/* major number of driver */
-    const char *major_name;	/* name of major driver */
-    int minor_shift;		/* number of times minor is shifted to
-				 * get real minor */
-    int max_p;			/* maximum partitions per device */
-    int max_nr;			/* maximum number of real devices */
-
-    void (*init) ();		/* Initialization called before we do our thing */
-    struct hd_struct *part;	/* partition table */
-    int *sizes;			/* device size in blocks, copied to blk_size[] */
-    int nr_real;		/* number of real devices */
-
-    void *real_devices;		/* internal use */
+    int major;                  /* major number of driver */
+    const char *major_name;     /* name of major driver */
+    int minor_shift;            /* number of times minor is shifted to get real minor */
+    int max_p;                  /* maximum partitions per device */
+    int max_nr;                 /* maximum number of real devices */
+    void (*init) ();            /* Initialization called before we do our thing */
+    struct hd_struct *part;     /* partition table */
+    int *sizes;                 /* device size in blocks, copied to blk_size[] */
+    int nr_hd;                  /* number of hard drives */
+    struct drive_infot *drive_info;
     struct gendisk *next;
 };
 
-extern struct gendisk *gendisk_head;	/* linked list of disks */
+extern struct drive_infot *last_drive;  /* set to last drivep-> used in read/write */
+extern unsigned char bios_drive_map[];  /* map drive to BIOS drivenum */
+
+extern struct gendisk *gendisk_head;    /* linked list of disks */
 
 #endif
